@@ -20,7 +20,7 @@ import 'package:pilipala/utils/utils.dart';
 class DynamicsController extends GetxController {
   int page = 1;
   String? offset = '';
-  RxList<DynamicItemModel> dynamicsList = [DynamicItemModel()].obs;
+  RxList<DynamicItemModel> dynamicsList = <DynamicItemModel>[].obs;
   Rx<DynamicsType> dynamicsType = DynamicsType.values[0].obs;
   RxString dynamicsTypeLabel = '全部'.obs;
   final ScrollController scrollController = ScrollController();
@@ -105,7 +105,7 @@ class DynamicsController extends GetxController {
 
   onSelectType(value) async {
     dynamicsType.value = filterTypeList[value]['value'];
-    dynamicsList.value = [DynamicItemModel()];
+    dynamicsList.value = <DynamicItemModel>[];
     page = 1;
     initialValue.value = value;
     await queryFollowDynamic();
@@ -249,8 +249,8 @@ class DynamicsController extends GetxController {
       return {'status': false, 'msg': '账号未登录'};
     }
     if (type == 'init') {
-      upData.value.upList = [];
-      upData.value.liveUsers = LiveUsers();
+      upData.value.upList = <UpItem>[];
+      upData.value.liveList = <LiveUserItem>[];
     }
     var res = await DynamicsHttp.followUp();
     if (res['status']) {
@@ -258,20 +258,23 @@ class DynamicsController extends GetxController {
       if (upData.value.upList!.isEmpty) {
         mid.value = -1;
       }
+      upData.value.upList!.insertAll(0, [
+        UpItem(face: '', uname: '全部动态', mid: -1),
+        UpItem(face: userInfo.face, uname: '我', mid: userInfo.mid),
+      ]);
     }
     return res;
   }
 
   onSelectUp(mid) async {
     dynamicsType.value = DynamicsType.values[0];
-    dynamicsList.value = [DynamicItemModel()];
+    dynamicsList.value = <DynamicItemModel>[];
     page = 1;
     queryFollowDynamic();
   }
 
   onRefresh() async {
     page = 1;
-    print('onRefresh');
     await queryFollowUp();
     await queryFollowDynamic();
   }
@@ -293,7 +296,7 @@ class DynamicsController extends GetxController {
     dynamicsType.value = DynamicsType.values[0];
     initialValue.value = 0;
     SmartDialog.showToast('还原默认加载');
-    dynamicsList.value = [DynamicItemModel()];
+    dynamicsList.value = <DynamicItemModel>[];
     queryFollowDynamic();
   }
 }
