@@ -7,8 +7,10 @@ import 'package:pilipala/models/video/play/quality.dart';
 import 'package:pilipala/pages/setting/widgets/select_dialog.dart';
 import 'package:pilipala/plugin/pl_player/index.dart';
 import 'package:pilipala/services/service_locator.dart';
+import 'package:pilipala/utils/global_data.dart';
 import 'package:pilipala/utils/storage.dart';
 
+import '../../models/live/quality.dart';
 import 'widgets/switch_item.dart';
 
 class PlaySetting extends StatefulWidget {
@@ -21,6 +23,7 @@ class PlaySetting extends StatefulWidget {
 class _PlaySettingState extends State<PlaySetting> {
   Box setting = GStrorage.setting;
   late dynamic defaultVideoQa;
+  late dynamic defaultLiveQa;
   late dynamic defaultAudioQa;
   late dynamic defaultDecode;
   late int defaultFullScreenMode;
@@ -31,6 +34,8 @@ class _PlaySettingState extends State<PlaySetting> {
     super.initState();
     defaultVideoQa = setting.get(SettingBoxKey.defaultVideoQa,
         defaultValue: VideoQuality.values.last.code);
+    defaultLiveQa = setting.get(SettingBoxKey.defaultLiveQa,
+        defaultValue: LiveQuality.values.last.code);
     defaultAudioQa = setting.get(SettingBoxKey.defaultAudioQa,
         defaultValue: AudioQuality.values.last.code);
     defaultDecode = setting.get(SettingBoxKey.defaultDecode,
@@ -72,6 +77,12 @@ class _PlaySettingState extends State<PlaySetting> {
             onTap: () => Get.toNamed('/playSpeedSet'),
             title: Text('倍速设置', style: titleStyle),
             subtitle: Text('设置视频播放速度', style: subTitleStyle),
+          ),
+          ListTile(
+            dense: false,
+            onTap: () => Get.toNamed('/playerGestureSet'),
+            title: Text('手势设置', style: titleStyle),
+            subtitle: Text('设置播放器手势', style: subTitleStyle),
           ),
           const SetSwitchItem(
             title: '开启1080P',
@@ -135,22 +146,24 @@ class _PlaySettingState extends State<PlaySetting> {
             defaultVal: false,
           ),
           const SetSwitchItem(
-            title: '双击快退/快进',
-            subTitle: '左侧双击快退，右侧双击快进',
-            setKey: SettingBoxKey.enableQuickDouble,
-            defaultVal: true,
-          ),
-          const SetSwitchItem(
             title: '弹幕开关',
             subTitle: '展示弹幕',
             setKey: SettingBoxKey.enableShowDanmaku,
             defaultVal: false,
           ),
+          SetSwitchItem(
+              title: '控制栏动画',
+              subTitle: '播放器控制栏显示动画效果',
+              setKey: SettingBoxKey.enablePlayerControlAnimation,
+              defaultVal: true,
+              callFn: (bool val) {
+                GlobalData().enablePlayerControlAnimation = val;
+              }),
           ListTile(
             dense: false,
-            title: Text('默认画质', style: titleStyle),
+            title: Text('默认视频画质', style: titleStyle),
             subtitle: Text(
-              '当前画质${VideoQualityCode.fromCode(defaultVideoQa)!.description!}',
+              '当前默认画质${VideoQualityCode.fromCode(defaultVideoQa)!.description!}',
               style: subTitleStyle,
             ),
             onTap: () async {
@@ -158,7 +171,7 @@ class _PlaySettingState extends State<PlaySetting> {
                 context: context,
                 builder: (context) {
                   return SelectDialog<int>(
-                      title: '默认画质',
+                      title: '默认视频画质',
                       value: defaultVideoQa,
                       values: VideoQuality.values.reversed.map((e) {
                         return {'title': e.description, 'value': e.code};
@@ -168,6 +181,32 @@ class _PlaySettingState extends State<PlaySetting> {
               if (result != null) {
                 defaultVideoQa = result;
                 setting.put(SettingBoxKey.defaultVideoQa, result);
+                setState(() {});
+              }
+            },
+          ),
+          ListTile(
+            dense: false,
+            title: Text('默认直播画质', style: titleStyle),
+            subtitle: Text(
+              '当前默认画质${LiveQualityCode.fromCode(defaultLiveQa)!.description!}',
+              style: subTitleStyle,
+            ),
+            onTap: () async {
+              int? result = await showDialog(
+                context: context,
+                builder: (context) {
+                  return SelectDialog<int>(
+                      title: '默认直播画质',
+                      value: defaultLiveQa,
+                      values: LiveQuality.values.reversed.map((e) {
+                        return {'title': e.description, 'value': e.code};
+                      }).toList());
+                },
+              );
+              if (result != null) {
+                defaultLiveQa = result;
+                setting.put(SettingBoxKey.defaultLiveQa, result);
                 setState(() {});
               }
             },
