@@ -15,16 +15,18 @@ import 'package:pilipala/pages/video/detail/widgets/ai_detail.dart';
 import 'package:pilipala/utils/feed_back.dart';
 import 'package:pilipala/utils/storage.dart';
 import 'package:pilipala/utils/utils.dart';
-
+import '../widgets/expandable_section.dart';
 import 'widgets/action_item.dart';
-import 'widgets/action_row_item.dart';
 import 'widgets/fav_panel.dart';
 import 'widgets/intro_detail.dart';
 import 'widgets/page.dart';
 import 'widgets/season.dart';
 
 class VideoIntroPanel extends StatefulWidget {
-  const VideoIntroPanel({super.key});
+  final String bvid;
+  final String? cid;
+
+  const VideoIntroPanel({super.key, required this.bvid, this.cid});
 
   @override
   State<VideoIntroPanel> createState() => _VideoIntroPanelState();
@@ -47,7 +49,8 @@ class _VideoIntroPanelState extends State<VideoIntroPanel>
 
     /// fix 全屏时参数丢失
     heroTag = Get.arguments['heroTag'];
-    videoIntroController = Get.put(VideoIntroController(), tag: heroTag);
+    videoIntroController =
+        Get.put(VideoIntroController(bvid: widget.bvid), tag: heroTag);
     _futureBuilderFuture = videoIntroController.queryVideoIntro();
     videoIntroController.videoDetail.listen((value) {
       videoDetail = value;
@@ -74,9 +77,9 @@ class _VideoIntroPanelState extends State<VideoIntroPanel>
             // 请求成功
             return Obx(
               () => VideoInfo(
-                loadingStatus: false,
                 videoDetail: videoIntroController.videoDetail.value,
                 heroTag: heroTag,
+                bvid: widget.bvid,
               ),
             );
           } else {
@@ -91,10 +94,13 @@ class _VideoIntroPanelState extends State<VideoIntroPanel>
             );
           }
         } else {
-          return VideoInfo(
-            loadingStatus: true,
-            videoDetail: videoDetail,
-            heroTag: heroTag,
+          return const SliverToBoxAdapter(
+            child: SizedBox(
+              height: 100,
+              child: Center(
+                child: CircularProgressIndicator(),
+              ),
+            ),
           );
         }
       },
@@ -103,31 +109,28 @@ class _VideoIntroPanelState extends State<VideoIntroPanel>
 }
 
 class VideoInfo extends StatefulWidget {
-  final bool loadingStatus;
   final VideoDetailData? videoDetail;
   final String? heroTag;
+  final String bvid;
 
-  const VideoInfo(
-      {Key? key, this.loadingStatus = false, this.videoDetail, this.heroTag})
-      : super(key: key);
+  const VideoInfo({
+    Key? key,
+    this.videoDetail,
+    this.heroTag,
+    required this.bvid,
+  }) : super(key: key);
 
   @override
   State<VideoInfo> createState() => _VideoInfoState();
 }
 
 class _VideoInfoState extends State<VideoInfo> with TickerProviderStateMixin {
-  // final String heroTag = Get.arguments['heroTag'];
   late String heroTag;
   late final VideoIntroController videoIntroController;
   late final VideoDetailController videoDetailCtr;
-  late final Map<dynamic, dynamic> videoItem;
-
   final Box<dynamic> localCache = GStrorage.localCache;
   final Box<dynamic> setting = GStrorage.setting;
   late double sheetHeight;
-
-  late final bool loadingStatus; // 加载状态
-
   late final dynamic owner;
   late final dynamic follower;
   late final dynamic followStatus;
@@ -135,6 +138,7 @@ class _VideoInfoState extends State<VideoInfo> with TickerProviderStateMixin {
   late String memberHeroTag;
   late bool enableAi;
   bool isProcessing = false;
+  RxBool isExpand = false.obs;
   void Function()? handleState(Future Function() action) {
     return isProcessing
         ? null
@@ -149,16 +153,13 @@ class _VideoInfoState extends State<VideoInfo> with TickerProviderStateMixin {
   void initState() {
     super.initState();
     heroTag = widget.heroTag!;
-    videoIntroController = Get.put(VideoIntroController(), tag: heroTag);
+    videoIntroController =
+        Get.put(VideoIntroController(bvid: widget.bvid), tag: heroTag);
     videoDetailCtr = Get.find<VideoDetailController>(tag: heroTag);
-    videoItem = videoIntroController.videoItem!;
     sheetHeight = localCache.get('sheetHeight');
 
-    loadingStatus = widget.loadingStatus;
-    owner = loadingStatus ? videoItem['owner'] : widget.videoDetail!.owner;
-    follower = loadingStatus
-        ? '-'
-        : Utils.numFormat(videoIntroController.userStat['follower']);
+    owner = widget.videoDetail!.owner;
+    follower = Utils.numFormat(videoIntroController.userStat['follower']);
     followStatus = videoIntroController.followStatus;
     enableAi = setting.get(SettingBoxKey.enableAi, defaultValue: true);
   }
@@ -212,29 +213,16 @@ class _VideoInfoState extends State<VideoInfo> with TickerProviderStateMixin {
 
   // 视频介绍
   showIntroDetail() {
-    if (loadingStatus) {
-      return;
-    }
     feedBack();
-    showBottomSheet(
-      context: context,
-      enableDrag: true,
-      builder: (BuildContext context) {
-        return IntroDetail(videoDetail: widget.videoDetail!);
-      },
-    );
+    isExpand.value = !(isExpand.value);
   }
 
   // 用户主页
   onPushMember() {
     feedBack();
-    mid = !loadingStatus
-        ? widget.videoDetail!.owner!.mid
-        : videoItem['owner'].mid;
+    mid = widget.videoDetail!.owner!.mid!;
     memberHeroTag = Utils.makeHeroTag(mid);
-    String face = !loadingStatus
-        ? widget.videoDetail!.owner!.face
-        : videoItem['owner'].face;
+    String face = widget.videoDetail!.owner!.face!;
     Get.toNamed('/member?mid=$mid',
         arguments: {'face': face, 'heroTag': memberHeroTag});
   }
@@ -256,223 +244,196 @@ class _VideoInfoState extends State<VideoInfo> with TickerProviderStateMixin {
     final Color outline = t.colorScheme.outline;
     return SliverPadding(
       padding: const EdgeInsets.only(
-          left: StyleString.safeSpace, right: StyleString.safeSpace, top: 10),
+        left: StyleString.safeSpace,
+        right: StyleString.safeSpace,
+        top: 16,
+      ),
       sliver: SliverToBoxAdapter(
-        child: !loadingStatus
-            ? Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  GestureDetector(
-                    behavior: HitTestBehavior.translucent,
-                    onTap: () => showIntroDetail(),
-                    child: Text(
-                      !loadingStatus
-                          ? widget.videoDetail!.title
-                          : videoItem['title'],
-                      style: const TextStyle(
-                        fontSize: 18,
-                        fontWeight: FontWeight.bold,
-                      ),
-                      maxLines: 2,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ),
-                  Stack(
+          child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          GestureDetector(
+            behavior: HitTestBehavior.translucent,
+            onTap: () => showIntroDetail(),
+            child: Text(
+              widget.videoDetail!.title!,
+              style: const TextStyle(
+                fontSize: 18,
+                fontWeight: FontWeight.bold,
+              ),
+              maxLines: 2,
+              overflow: TextOverflow.ellipsis,
+            ),
+          ),
+          Stack(
+            children: [
+              GestureDetector(
+                behavior: HitTestBehavior.translucent,
+                onTap: () => showIntroDetail(),
+                child: Padding(
+                  padding: const EdgeInsets.only(top: 7, bottom: 6),
+                  child: Row(
                     children: [
-                      GestureDetector(
-                        behavior: HitTestBehavior.translucent,
-                        onTap: () => showIntroDetail(),
-                        child: Padding(
-                          padding: const EdgeInsets.only(top: 7, bottom: 6),
-                          child: Row(
-                            children: [
-                              StatView(
-                                theme: 'gray',
-                                view: !loadingStatus
-                                    ? widget.videoDetail!.stat!.view
-                                    : videoItem['stat'].view,
-                                size: 'medium',
-                              ),
-                              const SizedBox(width: 10),
-                              StatDanMu(
-                                theme: 'gray',
-                                danmu: !loadingStatus
-                                    ? widget.videoDetail!.stat!.danmaku
-                                    : videoItem['stat'].danmaku,
-                                size: 'medium',
-                              ),
-                              const SizedBox(width: 10),
-                              Text(
-                                Utils.dateFormat(
-                                    !loadingStatus
-                                        ? widget.videoDetail!.pubdate
-                                        : videoItem['pubdate'],
-                                    formatType: 'detail'),
-                                style: TextStyle(
-                                  fontSize: 12,
-                                  color: t.colorScheme.outline,
-                                ),
-                              ),
-                              const SizedBox(width: 10),
-                              if (videoIntroController.isShowOnlineTotal)
-                                Obx(
-                                  () => Text(
-                                    '${videoIntroController.total.value}人在看',
-                                    style: TextStyle(
-                                      fontSize: 12,
-                                      color: t.colorScheme.outline,
-                                    ),
-                                  ),
-                                ),
-                            ],
-                          ),
+                      StatView(
+                        theme: 'gray',
+                        view: widget.videoDetail!.stat!.view,
+                        size: 'medium',
+                      ),
+                      const SizedBox(width: 10),
+                      StatDanMu(
+                        theme: 'gray',
+                        danmu: widget.videoDetail!.stat!.danmaku,
+                        size: 'medium',
+                      ),
+                      const SizedBox(width: 10),
+                      Text(
+                        Utils.dateFormat(widget.videoDetail!.pubdate,
+                            formatType: 'detail'),
+                        style: TextStyle(
+                          fontSize: 12,
+                          color: t.colorScheme.outline,
                         ),
                       ),
-                      if (enableAi)
-                        Positioned(
-                          right: 10,
-                          top: 6,
-                          child: GestureDetector(
-                            onTap: () async {
-                              final res =
-                                  await videoIntroController.aiConclusion();
-                              if (res['status']) {
-                                showAiBottomSheet();
-                              }
-                            },
-                            child:
-                                Image.asset('assets/images/ai.png', height: 22),
-                          ),
-                        )
-                    ],
-                  ),
-                  // 点赞收藏转发 布局样式1
-                  // SingleChildScrollView(
-                  //   padding: const EdgeInsets.only(top: 7, bottom: 7),
-                  //   scrollDirection: Axis.horizontal,
-                  //   child: actionRow(
-                  //     context,
-                  //     videoIntroController,
-                  //     videoDetailCtr,
-                  //   ),
-                  // ),
-                  // 点赞收藏转发 布局样式2
-                  actionGrid(context, videoIntroController),
-                  // 合集
-                  if (!loadingStatus &&
-                      widget.videoDetail!.ugcSeason != null) ...[
-                    Obx(
-                      () => SeasonPanel(
-                        ugcSeason: widget.videoDetail!.ugcSeason!,
-                        cid: videoIntroController.lastPlayCid.value != 0
-                            ? videoIntroController.lastPlayCid.value
-                            : widget.videoDetail!.pages!.first.cid,
-                        sheetHeight: sheetHeight,
-                        changeFuc: (bvid, cid, aid) => videoIntroController
-                            .changeSeasonOrbangu(bvid, cid, aid),
-                      ),
-                    )
-                  ],
-                  if (!loadingStatus &&
-                      widget.videoDetail!.pages != null &&
-                      widget.videoDetail!.pages!.length > 1) ...[
-                    Obx(() => PagesPanel(
-                          pages: widget.videoDetail!.pages!,
-                          cid: videoIntroController.lastPlayCid.value,
-                          sheetHeight: sheetHeight,
-                          changeFuc: (cid) =>
-                              videoIntroController.changeSeasonOrbangu(
-                                  videoIntroController.bvid, cid, null),
-                        ))
-                  ],
-                  GestureDetector(
-                    onTap: onPushMember,
-                    child: Container(
-                      padding: const EdgeInsets.symmetric(
-                          vertical: 12, horizontal: 4),
-                      child: Row(
-                        children: [
-                          NetworkImgLayer(
-                            type: 'avatar',
-                            src: loadingStatus
-                                ? owner.face
-                                : widget.videoDetail!.owner!.face,
-                            width: 34,
-                            height: 34,
-                            fadeInDuration: Duration.zero,
-                            fadeOutDuration: Duration.zero,
-                          ),
-                          const SizedBox(width: 10),
-                          Text(owner.name,
-                              style: const TextStyle(fontSize: 13)),
-                          const SizedBox(width: 6),
-                          Text(
-                            follower,
+                      const SizedBox(width: 10),
+                      if (videoIntroController.isShowOnlineTotal)
+                        Obx(
+                          () => Text(
+                            '${videoIntroController.total.value}人在看',
                             style: TextStyle(
-                              fontSize: t.textTheme.labelSmall!.fontSize,
-                              color: outline,
+                              fontSize: 12,
+                              color: t.colorScheme.outline,
                             ),
                           ),
-                          const Spacer(),
-                          Obx(() => AnimatedOpacity(
-                                opacity: loadingStatus ||
-                                        videoIntroController
-                                            .followStatus.isEmpty
-                                    ? 0
-                                    : 1,
-                                duration: const Duration(milliseconds: 50),
-                                child: SizedBox(
-                                  height: 32,
-                                  child: Obx(
-                                    () => videoIntroController
-                                            .followStatus.isNotEmpty
-                                        ? TextButton(
-                                            onPressed: videoIntroController
-                                                .actionRelationMod,
-                                            style: TextButton.styleFrom(
-                                              padding: const EdgeInsets.only(
-                                                  left: 8, right: 8),
-                                              foregroundColor:
-                                                  followStatus['attribute'] != 0
-                                                      ? outline
-                                                      : t.colorScheme.onPrimary,
-                                              backgroundColor:
-                                                  followStatus['attribute'] != 0
-                                                      ? t.colorScheme
-                                                          .onInverseSurface
-                                                      : t.colorScheme
-                                                          .primary, // 设置按钮背景色
-                                            ),
-                                            child: Text(
-                                              followStatus['attribute'] != 0
-                                                  ? '已关注'
-                                                  : '关注',
-                                              style: TextStyle(
-                                                  fontSize: t.textTheme
-                                                      .labelMedium!.fontSize),
-                                            ),
-                                          )
-                                        : ElevatedButton(
-                                            onPressed: videoIntroController
-                                                .actionRelationMod,
-                                            child: const Text('关注'),
-                                          ),
-                                  ),
-                                ),
-                              )),
-                        ],
-                      ),
-                    ),
+                        ),
+                    ],
                   ),
-                ],
-              )
-            : const SizedBox(
-                height: 100,
-                child: Center(
-                  child: CircularProgressIndicator(),
                 ),
               ),
-      ),
+              if (enableAi)
+                Positioned(
+                  right: 10,
+                  top: 6,
+                  child: GestureDetector(
+                    onTap: () async {
+                      final res = await videoIntroController.aiConclusion();
+                      if (res['status']) {
+                        showAiBottomSheet();
+                      }
+                    },
+                    child: Image.asset('assets/images/ai.png', height: 22),
+                  ),
+                )
+            ],
+          ),
+
+          /// 视频简介
+          Obx(
+            () => ExpandedSection(
+              expand: isExpand.value,
+              begin: 0,
+              end: 1,
+              child: IntroDetail(videoDetail: widget.videoDetail!),
+            ),
+          ),
+
+          /// 点赞收藏转发
+          actionGrid(context, videoIntroController),
+          // 合集
+          if (widget.videoDetail!.ugcSeason != null) ...[
+            Obx(
+              () => SeasonPanel(
+                ugcSeason: widget.videoDetail!.ugcSeason!,
+                cid: videoIntroController.lastPlayCid.value != 0
+                    ? videoIntroController.lastPlayCid.value
+                    : widget.videoDetail!.pages!.first.cid,
+                sheetHeight: sheetHeight,
+                changeFuc: (bvid, cid, aid) =>
+                    videoIntroController.changeSeasonOrbangu(bvid, cid, aid),
+              ),
+            )
+          ],
+          if (widget.videoDetail!.pages != null &&
+              widget.videoDetail!.pages!.length > 1) ...[
+            Obx(() => PagesPanel(
+                  pages: widget.videoDetail!.pages!,
+                  cid: videoIntroController.lastPlayCid.value,
+                  sheetHeight: sheetHeight,
+                  changeFuc: (cid) => videoIntroController.changeSeasonOrbangu(
+                      videoIntroController.bvid, cid, null),
+                ))
+          ],
+          GestureDetector(
+            onTap: onPushMember,
+            child: Container(
+              padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 4),
+              child: Row(
+                children: [
+                  NetworkImgLayer(
+                    type: 'avatar',
+                    src: widget.videoDetail!.owner!.face,
+                    width: 34,
+                    height: 34,
+                    fadeInDuration: Duration.zero,
+                    fadeOutDuration: Duration.zero,
+                  ),
+                  const SizedBox(width: 10),
+                  Text(owner.name, style: const TextStyle(fontSize: 13)),
+                  const SizedBox(width: 6),
+                  Text(
+                    follower,
+                    style: TextStyle(
+                      fontSize: t.textTheme.labelSmall!.fontSize,
+                      color: outline,
+                    ),
+                  ),
+                  const Spacer(),
+                  Obx(() => AnimatedOpacity(
+                        opacity:
+                            videoIntroController.followStatus.isEmpty ? 0 : 1,
+                        duration: const Duration(milliseconds: 50),
+                        child: SizedBox(
+                          height: 32,
+                          child: Obx(
+                            () => videoIntroController.followStatus.isNotEmpty
+                                ? TextButton(
+                                    onPressed:
+                                        videoIntroController.actionRelationMod,
+                                    style: TextButton.styleFrom(
+                                      padding: const EdgeInsets.only(
+                                          left: 8, right: 8),
+                                      foregroundColor:
+                                          followStatus['attribute'] != 0
+                                              ? outline
+                                              : t.colorScheme.onPrimary,
+                                      backgroundColor:
+                                          followStatus['attribute'] != 0
+                                              ? t.colorScheme.onInverseSurface
+                                              : t.colorScheme
+                                                  .primary, // 设置按钮背景色
+                                    ),
+                                    child: Text(
+                                      followStatus['attribute'] != 0
+                                          ? '已关注'
+                                          : '关注',
+                                      style: TextStyle(
+                                          fontSize: t
+                                              .textTheme.labelMedium!.fontSize),
+                                    ),
+                                  )
+                                : ElevatedButton(
+                                    onPressed:
+                                        videoIntroController.actionRelationMod,
+                                    child: const Text('关注'),
+                                  ),
+                          ),
+                        ),
+                      )),
+                ],
+              ),
+            ),
+          ),
+        ],
+      )),
     );
   }
 
@@ -483,6 +444,7 @@ class _VideoInfoState extends State<VideoInfo> with TickerProviderStateMixin {
         margin: const EdgeInsets.only(top: 6, bottom: 4),
         height: constraints.maxWidth / 5 * 0.8,
         child: GridView.count(
+          physics: const NeverScrollableScrollPhysics(),
           primary: false,
           padding: EdgeInsets.zero,
           crossAxisCount: 5,
@@ -494,117 +456,42 @@ class _VideoInfoState extends State<VideoInfo> with TickerProviderStateMixin {
                   selectIcon: const Icon(FontAwesomeIcons.solidThumbsUp),
                   onTap: handleState(videoIntroController.actionLikeVideo),
                   selectStatus: videoIntroController.hasLike.value,
-                  loadingStatus: loadingStatus,
-                  text: !loadingStatus
-                      ? widget.videoDetail!.stat!.like!.toString()
-                      : '-'),
-            ),
-            // ActionItem(
-            //     icon: const Icon(FontAwesomeIcons.clock),
-            //     onTap: () => videoIntroController.actionShareVideo(),
-            //     selectStatus: false,
-            //     loadingStatus: loadingStatus,
-            //     text: '稍后再看'),
-            Obx(
-              () => ActionItem(
-                  icon: const Icon(FontAwesomeIcons.b),
-                  selectIcon: const Icon(FontAwesomeIcons.b),
-                  onTap: handleState(videoIntroController.actionCoinVideo),
-                  selectStatus: videoIntroController.hasCoin.value,
-                  loadingStatus: loadingStatus,
-                  text: !loadingStatus
-                      ? widget.videoDetail!.stat!.coin!.toString()
-                      : '-'),
+                  text: widget.videoDetail!.stat!.like!.toString()),
             ),
             Obx(
               () => ActionItem(
-                  icon: const Icon(FontAwesomeIcons.star),
-                  selectIcon: const Icon(FontAwesomeIcons.solidStar),
-                  onTap: () => showFavBottomSheet(),
-                  onLongPress: () => showFavBottomSheet(type: 'longPress'),
-                  selectStatus: videoIntroController.hasFav.value,
-                  loadingStatus: loadingStatus,
-                  text: !loadingStatus
-                      ? widget.videoDetail!.stat!.favorite!.toString()
-                      : '-'),
+                icon: const Icon(FontAwesomeIcons.b),
+                selectIcon: const Icon(FontAwesomeIcons.b),
+                onTap: handleState(videoIntroController.actionCoinVideo),
+                selectStatus: videoIntroController.hasCoin.value,
+                text: widget.videoDetail!.stat!.coin!.toString(),
+              ),
+            ),
+            Obx(
+              () => ActionItem(
+                icon: const Icon(FontAwesomeIcons.star),
+                selectIcon: const Icon(FontAwesomeIcons.solidStar),
+                onTap: () => showFavBottomSheet(),
+                onLongPress: () => showFavBottomSheet(type: 'longPress'),
+                selectStatus: videoIntroController.hasFav.value,
+                text: widget.videoDetail!.stat!.favorite!.toString(),
+              ),
             ),
             ActionItem(
-                icon: const Icon(FontAwesomeIcons.comment),
-                onTap: () => videoDetailCtr.tabCtr.animateTo(1),
-                selectStatus: false,
-                loadingStatus: loadingStatus,
-                text: !loadingStatus
-                    ? widget.videoDetail!.stat!.reply!.toString()
-                    : '评论'),
+              icon: const Icon(FontAwesomeIcons.clock),
+              onTap: () => videoIntroController.actionShareVideo(),
+              selectStatus: false,
+              text: '稍后看',
+            ),
             ActionItem(
-                icon: const Icon(FontAwesomeIcons.shareFromSquare),
-                onTap: () => videoIntroController.actionShareVideo(),
-                selectStatus: false,
-                loadingStatus: loadingStatus,
-                text: '分享'),
+              icon: const Icon(FontAwesomeIcons.shareFromSquare),
+              onTap: () => videoIntroController.actionShareVideo(),
+              selectStatus: false,
+              text: '分享',
+            ),
           ],
         ),
       );
     });
-  }
-
-  Widget actionRow(BuildContext context, videoIntroController, videoDetailCtr) {
-    return Row(children: <Widget>[
-      Obx(
-        () => ActionRowItem(
-          icon: const Icon(FontAwesomeIcons.thumbsUp),
-          onTap: handleState(videoIntroController.actionLikeVideo),
-          selectStatus: videoIntroController.hasLike.value,
-          loadingStatus: loadingStatus,
-          text:
-              !loadingStatus ? widget.videoDetail!.stat!.like!.toString() : '-',
-        ),
-      ),
-      const SizedBox(width: 8),
-      Obx(
-        () => ActionRowItem(
-          icon: const Icon(FontAwesomeIcons.b),
-          onTap: handleState(videoIntroController.actionCoinVideo),
-          selectStatus: videoIntroController.hasCoin.value,
-          loadingStatus: loadingStatus,
-          text:
-              !loadingStatus ? widget.videoDetail!.stat!.coin!.toString() : '-',
-        ),
-      ),
-      const SizedBox(width: 8),
-      Obx(
-        () => ActionRowItem(
-          icon: const Icon(FontAwesomeIcons.heart),
-          onTap: () => showFavBottomSheet(),
-          onLongPress: () => showFavBottomSheet(type: 'longPress'),
-          selectStatus: videoIntroController.hasFav.value,
-          loadingStatus: loadingStatus,
-          text: !loadingStatus
-              ? widget.videoDetail!.stat!.favorite!.toString()
-              : '-',
-        ),
-      ),
-      const SizedBox(width: 8),
-      ActionRowItem(
-        icon: const Icon(FontAwesomeIcons.comment),
-        onTap: () {
-          videoDetailCtr.tabCtr.animateTo(1);
-        },
-        selectStatus: false,
-        loadingStatus: loadingStatus,
-        text:
-            !loadingStatus ? widget.videoDetail!.stat!.reply!.toString() : '-',
-      ),
-      const SizedBox(width: 8),
-      ActionRowItem(
-          icon: const Icon(FontAwesomeIcons.share),
-          onTap: () => videoIntroController.actionShareVideo(),
-          selectStatus: false,
-          loadingStatus: loadingStatus,
-          // text: !loadingStatus
-          //     ? widget.videoDetail!.stat!.share!.toString()
-          //     : '-',
-          text: '转发'),
-    ]);
   }
 }
